Move elements.json generation to src/generate-elements
