+++ src/analysis.ts
@@ -17,7 +17,7 @@ import * as jsonschema from 'jsonschema';
 import * as path from 'path';
 
 import {Descriptor, DocumentDescriptor, ElementDescriptor, InlineDocumentDescriptor, PropertyDescriptor} from './ast/ast';
-import {Elements, Attribute, Element, Event, Property} from './elements-metadata';
+import {Attribute, Element, Elements, Event, Property} from './elements-metadata';
 import {JsonDocument} from './json/json-document';
 import {Document} from './parser/document';
 import {camelCaseToKebab, trimLeft} from './utils';
@@ -61,235 +61,6 @@ export class Analysis {
       throw new Error(
           `Invalid schema_version in AnalyzedPackage. ` +
           `Expected 1.x.x, got ${analyzedPackage.schema_version}`);
-    }
-  }
-}
-
-export function generateElementMetadata(
-    analysis: Analysis, packagePath?: string): Elements {
-  const packageGatherer = new PackageGatherer();
-  const elementsGatherer = new ElementGatherer();
-  new AnalysisWalker(analysis.descriptors).walk([
-    packageGatherer, elementsGatherer
-  ]);
-
-  const packagesByDir: Map<string, Elements> =
-      packageGatherer.packagesByDir;
-  const elements = elementsGatherer.elements;
-  const elementsByPackageDir = new Map<string, Element[]>();
-
-  for (const element of elementsGatherer.elements) {
-    const matchingPackageDirs =
-        <string[]>Array.from(packagesByDir.keys())
-            .filter(dir => trimLeft(element.path, '/').startsWith(dir));
-    const longestMatchingPackageDir =
-        matchingPackageDirs.sort((a, b) => b.length - a.length)[0] || '';
-
-    if (longestMatchingPackageDir === '' && !packagesByDir.has('')) {
-      packagesByDir.set('', {schema_version: '1.0.0', elements: []});
-    }
-    packagesByDir.get(longestMatchingPackageDir)!.elements.push(element);
-    // We want element paths to be relative to the package directory.
-    element.path = trimLeft(
-        trimLeft(element.path, '/').substring(longestMatchingPackageDir.length),
-        '/');
-  }
-
-  return packagesByDir.get(packagePath);
-}
-
-function serializeElementDescriptor(
-    elementDescriptor: ElementDescriptor, path: string): Element|null {
-  const propChangeEvents: Event[] =
-      (elementDescriptor.properties || [])
-          .filter(p => p.notify)
-          .map(p => ({
-                 name: `${camelCaseToKebab(p.name)}-changed`,
-                 type: 'CustomEvent',
-                 description: `Fired when the \`${p.name}\` property changes.`
-               }));
-
-  if (!elementDescriptor.is) {
-    return null;
-  }
-  const properties = elementDescriptor.properties || [];
-  return {
-    tagname: elementDescriptor.is,
-    description: '',
-    superclass: 'HTMLElement',
-    path: path,
-    attributes: computeAttributesFromPropertyDescriptors(properties),
-    properties: properties.map(serializePropertyDescriptor),
-    styling: {
-      cssVariables: [],
-      selectors: [],
-    },
-    demos: (elementDescriptor.demos || []).map(d => d.path),
-    slots: [],
-    events: propChangeEvents,
-    metadata: {},
-  };
-}
-
-function serializePropertyDescriptor(p: PropertyDescriptor): Property {
-  const property: Property = {
-    name: p.name,
-    type: p.type || '?',
-    description: p.desc || '',
-  };
-  if (p.default) {
-    property.defaultValue = JSON.stringify(p.default);
-  }
-  const polymerMetadata: any = {};
-  const polymerMetadataFields = ['notify', 'observer', 'readOnly'];
-  for (const field of polymerMetadataFields) {
-    if (field in p) {
-      polymerMetadata[field] = p[field];
-    }
-  }
-  property.metadata = {polymer: polymerMetadata};
-  return property;
-}
-
-function computeAttributesFromPropertyDescriptors(props: PropertyDescriptor[]):
-    Attribute[] {
-  return props.map(prop => {
-    const attribute: Attribute = {
-      name: camelCaseToKebab(prop.name),
-      description: prop.desc || ''
-    };
-    if (prop.type) {
-      attribute.type = prop.type;
-    }
-    if (prop.default) {
-      attribute.type = prop.type;
-    }
-    return attribute;
-  });
-}
-
-class PackageGatherer implements AnalysisVisitor {
-  private _packageFiles: string[] = [];
-  packagesByDir = new Map<string, Elements>();
-  visitDocumentDescriptor(dd: DocumentDescriptor, path: Descriptor[]): void {
-    if (dd.document instanceof JsonDocument &&
-        (dd.document.url.endsWith('package.json') ||
-         dd.document.url.endsWith('bower.json'))) {
-      this._packageFiles.push(dd.document.url);
-    }
-  }
-
-  done() {
-    for (const packageFile of this._packageFiles) {
-      const dirname = path.dirname(packageFile);
-      if (!this.packagesByDir.has(dirname)) {
-        this.packagesByDir.set(
-            trimLeft(dirname, '/'), {schema_version: '1.0.0', elements: []});
-      }
-    }
-  }
-}
-
-class ElementGatherer implements AnalysisVisitor {
-  elements: Element[] = [];
-  private _elementPaths = new Map<ElementDescriptor, string>();
-  visitElement(element: ElementDescriptor, path: Descriptor[]): void {
-    let pathToElement: string|null = null;
-    for (const descriptor of path) {
-      if (descriptor instanceof DocumentDescriptor) {
-        pathToElement = descriptor.document.url;
-      }
-    }
-    if (!pathToElement) {
-      throw new Error(`Unable to determine path to element: ${element}`);
-    }
-    if (this._elementPaths.has(element)) {
-      if (this._elementPaths.get(element) !== pathToElement) {
-        throw new Error(
-            `Found element ${element} at distinct paths: ` +
-            `${pathToElement} and ${this._elementPaths.get(element)}`);
-      }
-    } else {
-      this._elementPaths.set(element, pathToElement);
-      const elem = serializeElementDescriptor(element, pathToElement);
-      if (elem) {
-        this.elements.push(elem);
-      }
-    }
-  }
-}
-
-abstract class AnalysisVisitor {
-  visitDocumentDescriptor?(dd: DocumentDescriptor, path: Descriptor[]): void;
-  visitInlineDocumentDescriptor?
-      (dd: InlineDocumentDescriptor<any>, path: Descriptor[]): void;
-  visitElement?(element: ElementDescriptor, path: Descriptor[]): void;
-  done?(): void;
-}
-
-class AnalysisWalker {
-  private _documents: DocumentDescriptor[];
-  private _path: Descriptor[] = [];
-  constructor(descriptors: DocumentDescriptor[]) {
-    this._documents = descriptors;
-  }
-  walk(visitors: AnalysisVisitor[]) {
-    this._path.length = 0;
-    for (const descriptor of this._documents) {
-      this._walkDocumentDescriptor(descriptor, visitors);
-    }
-    for (const visitor of visitors) {
-      if (visitor.done) {
-        visitor.done();
-      }
-    }
-  }
-
-  private _walkDocumentDescriptor(
-      dd: DocumentDescriptor, visitors: AnalysisVisitor[]) {
-    this._path.push(dd);
-
-    for (const visitor of visitors) {
-      if (visitor.visitDocumentDescriptor) {
-        visitor.visitDocumentDescriptor(dd, this._path);
-      }
-    }
-
-    for (const entity of dd.entities) {
-      this._walkEntity(entity, visitors);
-    }
-    for (const dependency of dd.dependencies) {
-      this._walkEntity(dependency, visitors);
-    }
-    this._path.pop();
-  }
-
-  private _walkInlineDocumentDescriptor(
-      dd: InlineDocumentDescriptor<any>, visitors: AnalysisVisitor[]) {
-    for (const visitor of visitors) {
-      if (visitor.visitInlineDocumentDescriptor) {
-        visitor.visitInlineDocumentDescriptor(dd, this._path);
-      }
-    }
-  }
-
-  private _walkEntity(entity: Descriptor, visitors: AnalysisVisitor[]) {
-    if (entity instanceof DocumentDescriptor) {
-      return this._walkDocumentDescriptor(entity, visitors);
-    } else if (entity instanceof InlineDocumentDescriptor) {
-      return this._walkInlineDocumentDescriptor(entity, visitors);
-    } else if (entity['type'] === 'element') {
-      return this._walkElement(<ElementDescriptor>entity, visitors);
-    }
-    throw new Error(`Unknown kind of descriptor: ${entity}`);
-  }
-
-  private _walkElement(
-      element: ElementDescriptor, visitors: AnalysisVisitor[]) {
-    for (const visitor of visitors) {
-      if (visitor.visitElement) {
-        visitor.visitElement(element, this._path);
-      }
     }
   }
 }
